tests process date function
